refactor: extract ANSI utilities to separate module to fix cyclic import
Create new task_logger/ansi.py module containing strip_ansi_codes() and
related ANSI patterns, removing the dependency on logger.py.

This resolves CodeQL cyclic import warnings:
- utils.py imports TaskLogger (TYPE_CHECKING only)
- logger.py imports strip_ansi_codes from utils
- New ansi.py has no dependencies on other task_logger modules

Changes:
- Create apps/backend/task_logger/ansi.py with strip_ansi_codes()
- Update __init__.py to import strip_ansi_codes directly from .ansi
- Update logger.py to import from .ansi instead of .utils (7 locations)
- Update capture.py to import from .ansi instead of .utils
- Update tests to import from task_logger.ansi directly
- Remove duplicate ANSI code patterns from utils.py

All 27 tests pass.
