+++ apps/backend/task_logger/logger.py
@@ -161,7 +161,7 @@ class TaskLogger:
         # Add phase start entry
         phase_message = message or f"Starting {phase_key} phase"
         if phase_message:
-            from .utils import strip_ansi_codes
+            from .ansi import strip_ansi_codes
 
             phase_message = strip_ansi_codes(phase_message)
         entry = LogEntry(
@@ -208,7 +208,7 @@ class TaskLogger:
             message or f"{'Completed' if success else 'Failed'} {phase_key} phase"
         )
         if phase_message:
-            from .utils import strip_ansi_codes
+            from .ansi import strip_ansi_codes
 
             phase_message = strip_ansi_codes(phase_message)
 
@@ -251,7 +251,7 @@ class TaskLogger:
         """
         # Sanitize content to remove ANSI escape codes before storage
         if content:
-            from .utils import strip_ansi_codes
+            from .ansi import strip_ansi_codes
 
             content = strip_ansi_codes(content)
 
@@ -324,7 +324,7 @@ class TaskLogger:
 
         # Sanitize content and detail before storage
         if content or detail:
-            from .utils import strip_ansi_codes
+            from .ansi import strip_ansi_codes
 
             if content:
                 content = strip_ansi_codes(content)
@@ -390,7 +390,7 @@ class TaskLogger:
 
         # Sanitize subphase before use
         if subphase:
-            from .utils import strip_ansi_codes
+            from .ansi import strip_ansi_codes
 
             subphase = strip_ansi_codes(subphase)
 
@@ -439,7 +439,7 @@ class TaskLogger:
 
         # Sanitize tool_input before use
         if tool_input:
-            from .utils import strip_ansi_codes
+            from .ansi import strip_ansi_codes
 
             tool_input = strip_ansi_codes(tool_input)
 
@@ -505,7 +505,7 @@ class TaskLogger:
             display_result = display_result[:297] + "..."
 
         # Sanitize all string fields before storage
-        from .utils import strip_ansi_codes
+        from .ansi import strip_ansi_codes
 
         if display_result:
             display_result = strip_ansi_codes(display_result)
